+++ src/backend/resolvers/mutations/organization/obm/relational-resolvers/obmAndPayer/replace.js
@@ -22,6 +22,14 @@ const getObmPayerMaterializationPipeline = (obmId) => [
     },
   },
   {
+    $lookup: {
+      from: 'books',
+      localField: 'bookIds',
+      foreignField: '_id',
+      as: 'books',
+    },
+  },
+  {
     $project: {
       payer: {
         $arrayElemAt: ['$payer', 0],
@@ -29,6 +37,8 @@ const getObmPayerMaterializationPipeline = (obmId) => [
       obm: {
         $arrayElemAt: ['$obm', 0],
       },
+      note: 1,
+      books: '$books.name',
     },
   },
   {
@@ -45,6 +55,8 @@ const getObmPayerMaterializationPipeline = (obmId) => [
         organization: 1,
         organizationTiny: 1,
       },
+      note: 1,
+      books: 1,
     },
   },
 ]
@@ -60,11 +72,15 @@ const connectObmAndPayer = async (
 
   const session = mongoClient.startSession()
 
-  const docsToInsert = connections.map(({ _id, payerId }) => ({
-    _id: _id ? ObjectId(_id) : ObjectId(),
-    payerId: ObjectId(payerId),
-    obmId,
-  }))
+  const docsToInsert = connections.map(
+    ({ _id, payerId, bookIds = [], note }) => ({
+      _id: _id ? ObjectId(_id) : ObjectId(),
+      payerId: ObjectId(payerId),
+      obmId,
+      bookIds: bookIds.map((bookId) => ObjectId(bookId)),
+      note,
+    })
+  )
 
   await session.withTransaction(async () => {
     // Step 1: replace all existing entries affiliated with given obm